Final tweaks to avoid redundancy
