+++ views/goals/vscode.ts
@@ -1,0 +1,2 @@
+const vscode = acquireVsCodeApi();
+export default vscode;
